# dev/drupal#200: Performance improvement