@demeritcowboy 2024-12-20
This might be ok, but one difference is that it doesn't resolve the same way, for example, drupal node aliases. But the question would be is anyone using CRM_Utils_System::url or similar to construct links like that. What I mean is for example if I have a node with an alias defined in drupal, e.g. mynode1 is the alias for /node/1, then before the patch:

`input: node/1 => output: /mynode1`

after the patch:

`input: node/1 => output: /node/1`

This might just be a minor difference for something like aliases. Off the top of my head I'm not able to think of an example with bigger consequences, or why it would be civi code trying to make that url instead of code living in a drupal module.

@jackrabbithanna @MegaphoneJon