@jaapjansma 2024-12-20
@demeritcowboy you are right but if you want to construct a url for a drupal node one would use the Drupal functionality for that and not `CRM_Utils_System::url` . So my assumption is that it would be safe to replace this. The change is in a large production site already for 4+ months and nothing seems broken.