@demeritcowboy 2024-12-24
Sorry I need to remove the merge-ready because the reason this is fast is because it completely bypasses the CivicrmPathProcessor, which was the focus in https://github.com/civicrm/civicrm-drupal-8/pull/88 and https://github.com/civicrm/civicrm-drupal-8/pull/98. So we need to understand more why that path processor is there.

It's possible it serves no purpose, but that doesn't feel right.