@MegaphoneJon 2024-12-24
CivicrmPathProcessor is important for paths that aren't in `civicrm_menu`.  This includes CiviReport instances and several import URLs, along with a handful of others.  See `CRM_Core_Menu::get()` starting [after the call to civicrm_menu](https://github.com/civicrm/civicrm-core/blob/master/CRM/Core/Menu.php#L638).