@demeritcowboy 2024-12-24
Actually I see there's two things - I was only looking at the generation of the link. When you actually visit the link, it still hits the pathprocessor so those links still work (without the processor you get page not found).

So I think it's now down to urls like civicrm/activity/fee/fi/fo/fum, where before the patch it would generate civicrm/activity/fee:fi:fo:fum and now it leaves it as civicrm/activity/fee/fi/fo/fum, but I don't know where this would matter. When you visit those links, regardless of how they're formatted, we're talking about links handled by some civi code. For it to matter there would need to be a(nother) drupal module that handles visits to some links starting with `/civicrm`.
If the link is _generated_ on the drupal side somewhere, either with slashes or colons, then this PR doesn't change anything because this PR doesn't affect when you visit a link.